Add hdf5 regional weather dataset test
This is skipped, because you only need to run it as a manual
sanity check after a significant refactor.

It checks that you create the exact same regional dataset as you had
done in a previous version of the code. It takes ~45 mins to run a full
extract so not suitable as a unit test